moved all fetch Logic to the parent component on site details, had the child call the parent methods
